* master: (331 commits)
  Adding blurblog permalink to story on blurblog. Adding story slug to permalink. Fixing story guid hashes.
  Re-enabling message property for facebook.
  Ignoring image for facebook shares.
  Commenting out facebook share message until they approve it.
  Cleaning up Facebook og description.
  Facebook using a website instead of article.
  Facebook being a PITA.
  Facebook wants ISO8601 datetime format.
  Facebook doesn't like going to other URLs.
  Facebook doesn't like going to other URLs.
  Dumb typo in new facebook og crap.
  Using blurblog permalink for new facebook shares.
  Fixing Facebook share to use fancy actions.
  Adding scroll to comments button to share bar. Thanks @afita.
  Turning off microformats for more errors.
  Fixing errors in timeouts to show the correct error. Also fixing microformats parsing issue and allow IPv6 URLs in enclosures to be ignored, fixing a bunch of feeds.
  Cleaning redis stories for 1% of all feed fetches.
  Refreshing feed on fetch.
  Fiddling with logging on dupe feeds.
  Adding bs facebook assets.
  ...

Conflicts:
	apps/rss_feeds/models.py
	settings.py
	utils/feed_fetcher.py

+++ media/ios/ASI/ASIDataCompressor.m
@@ -212,7 +212,7 @@
 
 + (NSError *)deflateErrorWithCode:(int)code
 {
-	return [NSError errorWithDomain:NetworkRequestErrorDomain code:ASICompressionError userInfo:[NSDictionary dictionaryWithObjectsAndKeys:[NSString stringWithFormat:@"Compression of data failed with code %hi",code],NSLocalizedDescriptionKey,nil]];
+	return [NSError errorWithDomain:NetworkRequestErrorDomain code:ASICompressionError userInfo:[NSDictionary dictionaryWithObjectsAndKeys:[NSString stringWithFormat:@"Compression of data failed with code %hi",(short)code],NSLocalizedDescriptionKey,nil]];
 }
 
 @synthesize streamReady;
